requirements.txt
pandas
numpy
scikit-learn
joblib
openpyxl
xlrd
xgboost
matplotlib 
seaborn
azure-ai-ml
azure-identity
python-dotenv
marshmallow==3.20.1
azureml-mlflow
mlflow